api/requirements.txt
alembic==1.4.2
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
coverage==5.2.1
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
nose2==0.9.2
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
threadpoolctl==2.1.0
urllib3==1.25.9
Werkzeug==1.0.1